+++ src/components/Organizer.vue
@@ -48,7 +48,12 @@
                 href="/faq.html"
                 class="external text-gray-500"
                 target="_blank"
-              >FAQ</a>
+              >FAQ</a> |
+              <a
+                href="https://osm.org/copyright"
+                class="external text-gray-500"
+                target="_blank"
+              >OSM</a>
             </p>
           </div>
           <div class="col-md-6 text-center text-md-right text-gray-400">