+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- added automatic latitude longitude fill provided by OpenStreetMap
 - added RSS source type
 
 ### Changed